requirements.txt
numpy==2.4.3
pandas==3.0.1
requests==2.31.0
python-dotenv==1.0.0
spotipy==2.23.0
google-api-python-client==2.97.0
matplotlib==3.10.8
seaborn==0.13.2
scipy==1.17.1
statsmodels==0.14.6
scikit-learn==1.8.0
pytest==9.0.3